requirements.txt
numpy
openai
faster_whisper
sounddevice
py3-tts
langchain
pyserial
inquirer